requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
certifi==2025.10.5
dnspython==2.8.0
frozenlist==1.8.0
idna==3.11
magic-filter==1.0.12
motor==3.7.1
multidict==6.7.0
propcache==0.4.1
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
pymongo==4.15.3
python-dotenv==1.2.1
typing-inspection==0.4.2
typing_extensions==4.15.0
yarl==1.22.0
paytechuz[fastapi]
fastapi==0.115.0
uvicorn==0.32.1
sqlalchemy==2.0.36
httpx==0.28.1
